src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/reattach-content-ancestor-expected.txt

CONSOLE ERROR: line 7: Uncaught TypeError: host.createShadowRoot is not a function
